+++ ChangeLog.md
@@ -10,6 +10,7 @@
     - audio.pl change accepts a new '--[no]video' option
     - audio.pl change becomes audio.pl normalize
     - Define new '--remove' option for audio.pl normalize
+    - Define new '--check-album-path' and '--check-track-path' options for audio.pl list
 
 ### 4.19.0